C-style codes added, autotools integrated.
