+++ BUILD
@@ -1,0 +1,6 @@
+make distclean
+aclocal
+autoconf
+automake -i --add-missing
+./configure
+make
