--- CMakeLists.txt
@@ -1,11 +1,0 @@
-cmake_minimum_required (VERSION 3.16)
-project (deccenc)
-
-
-add_subdirectory (traceme)
-add_subdirectory (component)
-add_subdirectory (logger)
-add_subdirectory (util)
-add_subdirectory (media)
-
-add_subdirectory (test)
