+++ Makefile.am
@@ -1,0 +1,2 @@
+SUBDIRS = traceme common util component media
+CLEANFILES = *.a *.o Makefile.in Makefile
