--- build.sh
@@ -1,7 +1,0 @@
-
-#rm -rf build
-mkdir build
-cd build
-cmake ..
-cmake --build . --config Release
-cd -
