+++ common/Makefile.am
@@ -1,0 +1,1 @@
+CLEANFILES = *.a *.o Makefile.in Makefile
