+++ component/Makefile.am
@@ -1,0 +1,4 @@
+noinst_LIBRARIES = libany_data.a
+libany_data_a_CPPFLAGS = -I../common/
+libany_data_a_SOURCES = any_data.c
+CLEANFILES = *.a *.o Makefile.in Makefile
