+++ component/any_data.c
@@ -1,0 +1,47 @@
+// any_data.c
+// Foundational class.
+
+/**
+MIT License
+
+Copyright (c) 2022 Sourav Batabyal
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+
+#include <string.h>
+#include <stdlib.h>
+#include "any_data.h"
+#include "memutil.h"
+
+/**
+  * Set data name; this is the unique identifier of the data.
+  */
+RType set_name(AnyData* data, const char* const name, size_t name_len)
+{
+    if (IS_NULL(data) || IS_NULL(name))
+        return RT_NULL_Pointer;
+    if (data->name_ != NULL)
+    {
+        free(data->name_);
+        data->name_ = NULL;
+    }
+    void* dst = memcpy((void *)data->name_, (void *)name, name_len);
+    IS_NULL((char*)dst)? RT_Error : RT_Success;
+}
