--- component/inc/singular_component.h
@@ -1,45 +1,0 @@
-// singular_component.h
-// Foundational class.
-
-/**
-MIT License
-
-Copyright (c) 2022 Sourav Batabyal
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-*/
-
-#pragma once
-
-#ifndef _SINGULAR_COMPONENT_H_
-#define _SINGULAR_COMPONENT_H_
-
-#include <iostream>
-#include "component.h"
-
-namespace deccenc {
-class SingularComponent : public Component {
-public:
-	SingularComponent(const char* const name);
-	virtual ~SingularComponent();
-	static SingularComponent* ref_;
-};
-}
-
-#endif
