+++ configure.ac
@@ -1,14 +1,29 @@
+AC_PREREQ([2.69])
 AC_INIT([deccenc], [1.0])
-AM_INIT_AUTOMAKE([-Wall -Werror foreign])
+AM_INIT_AUTOMAKE([foreign -Wall -Werror])
+
+# Checks for programs.
+AM_PROG_AR
+AC_PROG_CC
 AC_PROG_CXX
+AC_PROG_RANLIB
+
+# Checks for libraries.
+
+# Checks for header files.
+
+# Checks for typedefs, structures, and compiler characteristics.
+
+# Checks for library functions.
+
 AC_CONFIG_FILES([
-        CMakeLists.txt
-        traceme/CMakeLists.txt
-        component/CMakeLists.txt
-        util/CMakeLists.txt
-        media/CMakeLists.txt
-        logger/CMakeLists.txt
-        test/CMakeLists.txt
-        test/media_test/CMakeLists.txt
+        Makefile 
+        traceme/Makefile
+        common/Makefile
+        util/Makefile
+        util/time/Makefile
+        component/Makefile
+        media/Makefile
 ])
+
 AC_OUTPUT
