+++ media/Makefile.am
@@ -1,0 +1,4 @@
+noinst_LIBRARIES = libmedia.a
+libmedia_a_CPPFLAGS = -I../common/ -I../compoment/
+libmedia_a_SOURCES = media.c
+CLEANFILES = *.a *.o Makefile.in Makefile
