--- test/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory("logger_test")
-add_subdirectory("media_test")
