--- test/logger_test/CMakeLists.txt
@@ -1,7 +1,0 @@
-add_executable(logger_test
-	src/main.cpp
-)
-target_link_libraries(logger_test
-	PUBLIC logger
-	PUBLIC component
-	PUBLIC traceme)
