--- test/media_test/CMakeLists.txt
@@ -1,8 +1,0 @@
-add_executable(media_test
-	src/main.cpp
-)
-target_link_libraries(media_test
-	PUBLIC logger
-	PUBLIC component
-	PUBLIC traceme
-	PUBLIC media)
