--- traceme/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_library(traceme 
-	src/traceme.cpp
-)
-target_include_directories(traceme
-	PUBLIC inc)
