--- util/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_library(util 
-	random/src/random.cpp
-)
-target_include_directories(util
-	PUBLIC random/inc)
