+++ util/Makefile.am
@@ -1,0 +1,2 @@
+SUBDIRS = time
+CLEANFILES = *.a *.o Makefile.in Makefile
