+++ util/time/Makefile.am
@@ -1,0 +1,3 @@
+noinst_LIBRARIES = libt_utility.a
+libt_utility_a_SOURCES = t_utility.c
+CLEANFILES = *.a *.o Makefile.in Makefile
